Crate iron_pipeline

Simple pipelined request handler for the Iron framework. Inspired by similar APIs used in the new ASP.NET Core 1.0 framework, which in turn is inspired by Microsoft's OWIN implementation, node.js and other web frameworks.

About

Under iron-pipeline, every request is sent through a daisy chain of middlewares, each of which may optionally:

Create and return a response

Modify the request

Delegate to the next middleware in the pipeline

Modify the response created by another middleware

Unlike Chain, middleware is always executed in the order in which it was registered.

Examples

This example introduces two helper middlewares: Handle and HandleNext.

use iron::prelude::*;
use iron::status;
use iron_pipeline::prelude::*;
 

fn main() {
    let mut pipeline = Pipeline::new();
     
    // "Middleware" example
    pipeline.add(HandleNext(|req, next| {
        log_request(req);
        let res = next.process(req);
        log_response(&res);
        res
    }));
     
    // "Handler" example
    pipeline.add(Handle(|req| {
        Ok(Response::with((
            status::Ok,
            "Hello from iron-pipeline"
        )))
    }));

    Iron::new(pipeline); // etc...
}

Usage

This crate may be used by adding iron-pipeline to the dependencies in your project's Cargo.toml:

[dependencies]
iron-pipeline = { git = "https://github.com/deadalusai/iron-pipeline" }

and the following to your crate root:

extern crate iron_pipeline;

Modules

middleware
Middleware for use in a Pipeline.

prelude
Includes the Pipeline type and all middleware types in the middleware module.

Structs

Pipeline
Iron middleware for implementing a simple forward-only pipeline. When a request is received each middleware is invoked in the order in which they were registered.

PipelineNext
Handle used to invoke the next handler in a pipeline

Enums

Error
Errors which may be raised by the Pipeline itself

Traits

PipelineMiddleware
Trait which defines middleware within a pipeline. Implementors of this trait must call next.handle(...) in order to pass control to the next middleware in the pipeline.