Struct iron_pipeline::middleware::fork::Fork

pub struct Fork<P>(_, _);

Middleware which optionally delegates to a sub pipeline based on some predicate applied to each request.

Methods

impl Fork<()>

fn when<P, B>(predicate: P, pipeline_builder: B) -> Fork<ForkOnFn<P>> where B: FnOnce(&mut Pipeline), P: Fn(&Request) -> bool + Send + Sync

Construct a new pipeline fork. The predicate is executed on every request and determines whether to delegate to the sub pipeline. The pipeline_builder is used to construct the sub pipeline, and is executed immediately.

Examples

A sub pipeline which handles all "Post" requests:

pipeline.add(Fork::when(|req| req.method == Method::Post, |sub_pipeline| {
    sub_pipeline.add(Handle(|req| {
        Ok(Response::with("Hello from iron-pipeline"))
    }));
}))

fn when_path<P, B>(path: P, pipeline_builder: B) -> Fork<ForkOnPath> where B: FnOnce(&mut Pipeline), P: AsRef<str>

Construct a new pipeline fork. The path is compared against the url path on every request and determines whether to delegate to the sub pipeline. The pipeline_builder is used to construct the sub pipeline, and is executed immediately.

Examples

A sub pipeline which handles all requests to "/api/v2":

pipeline.add(Fork::when_path("/api/v2", |sub_pipeline| {
    sub_pipeline.add(Handle(|req| {
        Ok(Response::with("Hello from iron-pipeline"))
    }));
}))

Panics

Panics when passed an invalid path string. Path should be of the form /hello/world.