Struct iron_pipeline::middleware::handle::HandleNext

pub struct HandleNext<F>(pub F) where F: Send + Sync + Fn(&mut Request, PipelineNext) -> IronResult<Response>;

Container for a pipeline middleware function which may optionally invoke the next middleware in the pipeline via next.process(req), or create a response itself.

Examples

pipeline.add(HandleNext(|req: &mut Request, next: PipelineNext| {
    log_request(req);
    let res = next.process(req);
    log_response(&res);
    res
}))