Struct iron_pipeline::Pipeline

pub struct Pipeline { /* fields omitted */ }

Iron middleware for implementing a simple forward-only pipeline. When a request is received each middleware is invoked in the order in which they were registered.

Each middleware may modify the request at will. It may then complete the request immediately or invoke to the next middleware in the pipeline.

Examples

let mut pipeline = Pipeline::new();
pipeline.add(MyCustomRequestPreprocessor);
pipeline.add(Handle(|req| Ok(Response::with((status::Ok, "Hello, world")))));
Iron::new(pipeline); // Etc...

Methods

impl Pipeline

fn new() -> Pipeline

Construct a new, empty request pipeline.

fn add<P>(&mut self, handler: P) where P: PipelineMiddleware + 'static

Append a middleware to the end of the pipeline